# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API ",
        "description": "Listado de los productos extras",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/Pulsera": {
            "get": {
                "tags": [
                    "Pulseras"
                ],
                "summary": "Consultar todas las pulseras",
                "description": "Retorna todas las pulseras disponibles",
                "operationId": "87a1c9beed1b94d8233ac47a70af6c9c",
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Pulsera de cuero"
                                                    },
                                                    "color": {
                                                        "type": "string",
                                                        "example": "negro"
                                                    },
                                                    "precio": {
                                                        "type": "number",
                                                        "example": "20"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pulseras"
                ],
                "summary": "Crear una nueva Pulsera",
                "description": "Crea una nueva Pulsera en el sistema.",
                "operationId": "b0a1c02cf01b89be720ed8875620f52b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Nombre",
                                    "Material",
                                    "Tamaño",
                                    "Precio",
                                    "Stock"
                                ],
                                "properties": {
                                    "Nombre": {
                                        "type": "string",
                                        "example": "Pulsera de cuero"
                                    },
                                    "Material": {
                                        "type": "string",
                                        "example": "Cuero"
                                    },
                                    "Tamaño": {
                                        "type": "string",
                                        "example": "M"
                                    },
                                    "Precio": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 25.99
                                    },
                                    "Stock": {
                                        "type": "integer",
                                        "example": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pulsera creada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "Nombre": {
                                            "type": "string",
                                            "example": "Pulsera de cuero"
                                        },
                                        "Material": {
                                            "type": "string",
                                            "example": "Cuero"
                                        },
                                        "Tamaño": {
                                            "type": "string",
                                            "example": "M"
                                        },
                                        "Precio": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 25.99
                                        },
                                        "Stock": {
                                            "type": "integer",
                                            "example": 50
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud incorrecta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Pulsera/{id}": {
            "put": {
                "tags": [
                    "Pulseras"
                ],
                "summary": "Actualizar una Pulsera existente",
                "description": "Actualiza los detalles de una Pulsera con el ID proporcionado.",
                "operationId": "160ecb7b514d5fb08e8940318c92ffcc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Nombre",
                                    "Material",
                                    "Tamaño",
                                    "Precio",
                                    "Stock"
                                ],
                                "properties": {
                                    "Nombre": {
                                        "type": "string",
                                        "example": "Pulsera de plata"
                                    },
                                    "Material": {
                                        "type": "string",
                                        "example": "Plata"
                                    },
                                    "Tamaño": {
                                        "type": "string",
                                        "example": "L"
                                    },
                                    "Precio": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 30.99
                                    },
                                    "Stock": {
                                        "type": "integer",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pulsera actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "Nombre": {
                                            "type": "string",
                                            "example": "Pulsera de plata"
                                        },
                                        "Material": {
                                            "type": "string",
                                            "example": "Plata"
                                        },
                                        "Tamaño": {
                                            "type": "string",
                                            "example": "L"
                                        },
                                        "Precio": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 30.99
                                        },
                                        "Stock": {
                                            "type": "integer",
                                            "example": 100
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontró la Pulsera",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Pulsera] #id"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pulseras"
                ],
                "summary": "Eliminar una Pulsera",
                "description": "Elimina una Pulsera del sistema usando el ID proporcionado.",
                "operationId": "b78d8d78a0cd40c32ae7ce69f85c7a54",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Pulsera eliminada correctamente"
                    },
                    "404": {
                        "description": "No se encontró la Pulsera",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Pulsera] #id"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Sanctum bearer token",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "bearer_token": []
        }
    ]
}